Func refactoring

+++ TopSolutions.ConsoleApp/GeneralCSharp/Func/Calculator.cs
@@ -8,9 +8,9 @@ namespace TopSolutions.ConsoleApp.GeneralCSharp.Func
 {
     public class Calculator
     {
-        public int Add(int a, int b) => a + b;
+        public double Add(int a, int b) => a + b;
         public int Subtract(int a, int b) => a - b;
-        public int Multiply(int a, int b) => a * b;
+        public double Multiply(int a, int b) => a * b;
         public int Divide(int a, int b) => a / b;
     }
 }
